[refactor] Subject Info enum 을 view 에서 viewModel로 이동
